Moving a JSONPath query test to the relevant package.
